+++ src/main/java/zaftnotameni/creatania/registry/CreataniaRegistrate.java
@@ -1,4 +1,5 @@
 package zaftnotameni.creatania.registry;
+
 import com.simibubi.create.foundation.data.CreateRegistrate;
 import com.tterrag.registrate.builders.FluidBuilder;
 import com.tterrag.registrate.providers.DataGenContext;
@@ -13,17 +14,21 @@ import net.minecraft.world.level.material.Fluid;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fluids.FluidAttributes;
 import net.minecraftforge.fluids.ForgeFlowingFluid;
+
 public class CreataniaRegistrate extends CreateRegistrate {
+
   public CreataniaRegistrate(String modid) {
     super(modid);
   }
+
   public static CreataniaRegistrate forMod(String id) { return new CreataniaRegistrate(id); }
+
   @Override
   public CreateRegistrate registerEventListeners(IEventBus bus) { return super.registerEventListeners(bus); }
 
-
   public static <I extends BlockItem> NonNullBiConsumer<DataGenContext<Item, I>, RegistrateItemModelProvider> sameAsBlockItemModel(
-    String... folders) {
+    String... folders
+  ) {
     return (c, p) -> {
       String path = "block";
       for (String string : folders)
@@ -31,34 +36,35 @@ public class CreataniaRegistrate extends CreateRegistrate {
       p.withExistingParent("item/" + c.getName(), p.modLoc(path));
     };
   }
+
   public FluidBuilder<ForgeFlowingFluid.Flowing, CreateRegistrate> standardFluid(String name) {
     return fluid(name, Index.resource("fluid/" + name + "_still"), Index.resource("fluid/" + name + "_flow"));
   }
 
-  public FluidBuilder<ForgeFlowingFluid.Flowing, CreateRegistrate> waterLikeFluid(String name,
-                                                                                     NonNullBiFunction<FluidAttributes.Builder, Fluid, FluidAttributes> attributesFactory) {
-    return fluid(name,
-      Fluids.WATER_STILL_RL,
-      Fluids.WATER_FLOWING_RL,
+  public FluidBuilder<ForgeFlowingFluid.Flowing, CreateRegistrate> waterLikeFluid(
+    String name,
+    NonNullBiFunction<FluidAttributes.Builder, Fluid, FluidAttributes> attributesFactory
+  ) {
+    return fluid(
+      name,
+      Fluids.MANA_STILL_RL,
+      Fluids.MANA_FLOWING_RL,
       attributesFactory,
-      Fluids.CreataniaFlowingFluidFlowing::new);
+      Fluids.CreataniaFlowingFluidFlowing::new
+    );
   }
-  public FluidBuilder<ForgeFlowingFluid.Flowing, CreateRegistrate> lavaLikeFluid(String name,
-                                                                                     NonNullBiFunction<FluidAttributes.Builder, Fluid, FluidAttributes> attributesFactory) {
-    return fluid(name,
-      Fluids.LAVA_STILL_RL,
-      Fluids.LAVA_FLOWING_RL,
-      attributesFactory,
-      Fluids.CreataniaFlowingFluidFlowing::new);
 
-  }
-  public FluidBuilder<ForgeFlowingFluid.Flowing, CreateRegistrate> texturedFluid(String name,
-                                                                                 NonNullBiFunction<FluidAttributes.Builder, Fluid, FluidAttributes> attributesFactory) {
-    return fluid(name,
-      Index.resource("fluid/" + name + "_still"),
-      Index.resource("fluid/" + name + "_flow"),
+  public FluidBuilder<ForgeFlowingFluid.Flowing, CreateRegistrate> lavaLikeFluid(
+    String name,
+    NonNullBiFunction<FluidAttributes.Builder, Fluid, FluidAttributes> attributesFactory
+  ) {
+    return fluid(
+      name,
+      Fluids.MOLTEN_STILL_RL,
+      Fluids.MOLTEN_FLOWING_RL,
       attributesFactory,
-    Fluids.CreataniaFlowingFluidFlowing::new);
+      Fluids.CreataniaFlowingFluidFlowing::new
+    );
 
   }
 
@@ -66,4 +72,5 @@ public class CreataniaRegistrate extends CreateRegistrate {
     super.creativeModeTab(creataniaCreativeTab);
     return this;
   }
+
 }