+++ src/main/java/zaftnotameni/creatania/registry/Fluids.java
@@ -1,4 +1,5 @@
 package zaftnotameni.creatania.registry;
+
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.tterrag.registrate.util.entry.FluidEntry;
@@ -27,6 +28,7 @@ import static zaftnotameni.creatania.util.Humanity.keyResource;
 import static zaftnotameni.creatania.util.Humanity.lang;
 
 public class Fluids {
+
   public static final ResourceLocation WATER_STILL_RL = new ResourceLocation("block/water_still");
   public static final ResourceLocation WATER_FLOWING_RL = new ResourceLocation("block/water_flow");
   public static final ResourceLocation WATER_OVERLAY_RL = new ResourceLocation("block/water_overlay");
@@ -34,6 +36,11 @@ public class Fluids {
   public static final ResourceLocation LAVA_FLOWING_RL = new ResourceLocation("block/lava_flow");
   public static final ResourceLocation LAVA_OVERLAY_RL = new ResourceLocation("block/water_overlay");
 
+  public static final ResourceLocation MOLTEN_STILL_RL = Index.resource("fluid/molten_still");
+  public static final ResourceLocation MOLTEN_FLOWING_RL = Index.resource("fluid/molten_flow");
+
+  public static final ResourceLocation MANA_STILL_RL = Index.resource("fluid/mana_still");
+  public static final ResourceLocation MANA_FLOWING_RL = Index.resource("fluid/mana_flow");
 
   // molten vanilla
   public static final FluidEntry<ForgeFlowingFluid.Flowing> MOLTEN_GOLD = registerMoltenFluid("gold", 0xffffff00);
@@ -50,18 +57,27 @@ public class Fluids {
   public static final FluidEntry<ForgeFlowingFluid.Flowing> MOLTEN_GAIA = registerMoltenFluid("gaia", 0xffffffff);
   // mana
   public static final FluidEntry<ForgeFlowingFluid.Flowing> PURE_MANA = registerManaFluid("pure_mana", 0xff11aaff, Tags.Fluids.PURE_MANA);
-  public static final FluidEntry<ForgeFlowingFluid.Flowing> CORRUPT_MANA = registerManaFluid("corrupt_mana", 0xff440044,Tags.Fluids.CORRUPT_MANA);
-  public static final FluidEntry<ForgeFlowingFluid.Flowing> REAL_MANA = registerManaFluid("real_mana", 0xff44ffff,Tags.Fluids.REAL_MANA);
+  public static final FluidEntry<ForgeFlowingFluid.Flowing> CORRUPT_MANA = registerManaFluid("corrupt_mana", 0xff440044, Tags.Fluids.CORRUPT_MANA);
+  public static final FluidEntry<ForgeFlowingFluid.Flowing> REAL_MANA = registerManaFluid("real_mana", 0xff44ffff, Tags.Fluids.REAL_MANA);
 
   public static JsonElement provideLangEntries() {
     var json = new JsonObject();
-    Index.all().getAll(Fluid.class).forEach(
-      entry -> json.addProperty("fluid." + keyResource(entry.getId()),
-        Humanity.slashes(lang.get().getAutomaticName (entry))));
+    Index
+      .all()
+      .getAll(Fluid.class)
+      .forEach(
+        entry -> json.addProperty(
+          "fluid." + keyResource(entry.getId()),
+          Humanity.slashes(lang
+                             .get()
+                             .getAutomaticName(entry))
+        ));
     return json;
   }
+
   public static FluidAttributes.Builder defaultMolten(FluidAttributes.Builder in, int color) {
-    return in.density(15)
+    return in
+      .density(15)
       .luminosity(15)
       .viscosity(10)
       .temperature(9000)
@@ -69,55 +85,65 @@ public class Fluids {
       .overlay(LAVA_OVERLAY_RL)
       .color(color);
   }
+
   public static void register(IEventBus bus) {
     Log.LOGGER.debug("register fluids");
   }
 
   public static FluidEntry<ForgeFlowingFluid.Flowing> registerManaFluid(String name, int color, TagKey<Fluid> tag) {
-    return Index.all().waterLikeFluid(name, Colored.from(color))
+    return Index
+      .all()
+      .waterLikeFluid(name, Colored.from(color))
       .lang("Liquid " + StringUtils.capitalize(name))
       .attributes(b -> defaultMolten(b, color).sound(HONEY_DRINK, HONEY_BLOCK_PLACE))
-      .properties(p -> p.tickRate(1 * 20)
+      .properties(p -> p
+        .tickRate(1 * 20)
         .levelDecreasePerBlock(2)
         .slopeFindDistance(2)
         .explosionResistance(100f))
-      .tag(Tags.Fluids.ALL_MANA, tag)
+      .tag(Tags.Fluids.ALL_MANA, tag, Tags.Fluids.minecraftTag("water"))
       .source(CreataniaFlowingFluidSource::new)
+      .tag(Tags.Fluids.minecraftTag("water"))
       .bucket()
-      .model((ctx, prov) -> prov.generated(ctx::getEntry, new ResourceLocation("minecraft", "item/lava_bucket")))
-      .tag(Tags.Items.tag("buckets/molten/" + name))
+      .model((ctx, prov) -> prov.generated(ctx::getEntry, Index.resource("fluid/mana_bucket")))
+      .tag(Tags.Items.tag("buckets/mana/" + name), Tags.Items.minecraftTag("water"))
       .build()
       .block()
+      .tag(Tags.Blocks.minecraftTag("water"))
       .build()
       .register();
   }
 
   public static FluidEntry<ForgeFlowingFluid.Flowing> registerMoltenFluid(String name, int color) {
-    return Index.all().lavaLikeFluid("molten_" + name, Colored.from(color))
+    return Index
+      .all()
+      .lavaLikeFluid("molten_" + name, Colored.from(color))
       .lang("Molten " + StringUtils.capitalize(name))
       .attributes(b -> defaultMolten(b, color))
-      .properties(p -> p.tickRate(5 * 20)
+      .properties(p -> p
+        .tickRate(5 * 20)
         .levelDecreasePerBlock(3)
         .slopeFindDistance(2)
         .explosionResistance(100f))
-      .tag(Tags.Fluids.MOLTEN)
+      .tag(Tags.Fluids.MOLTEN, Tags.Fluids.minecraftTag("water"))
       .source(CreataniaFlowingFluidSource::new)
+      .tag(Tags.Fluids.minecraftTag("water"))
       .bucket()
-      .model((ctx, prov) -> prov.generated(ctx::getEntry, new ResourceLocation("minecraft", "item/lava_bucket")))
-      .tag(Tags.Items.tag("buckets/molten/" + name))
+      .model((ctx, prov) -> prov.generated(ctx::getEntry, Index.resource("fluid/molten_bucket")))
+      .tag(Tags.Items.tag("buckets/molten/" + name), Tags.Items.minecraftTag("water"))
       .build()
       .block()
+      .tag(Tags.Blocks.minecraftTag("water"))
       .build()
       .register();
   }
 
-
-
-
-
   public static class Colored extends FluidAttributes {
+
     public Colored(Builder builder, Fluid fluid) { super(builder, fluid); }
+
     public int color = 0x00ffffff;
+
     public static NonNullBiFunction<Builder, Fluid, FluidAttributes> from(int color) {
       return (b, f) -> {
         var self = new Colored(b, f);
@@ -125,28 +151,36 @@ public class Fluids {
         return self;
       };
     }
+
     @Override
     public int getColor(BlockAndTintGetter world, BlockPos pos) {
       return color;
     }
+
   }
 
   public static class CreataniaFlowingFluidFlowing extends ForgeFlowingFluid.Flowing {
+
     public CreataniaFlowingFluidFlowing(Properties properties) { super(properties); }
+
     @Override
     protected void spreadTo(LevelAccessor pLevel, BlockPos pPos, BlockState pBlockState, Direction pDirection, FluidState pFluidState) {
-      if (!specialCobblegenSpread(pLevel, pPos, pBlockState, pDirection, pFluidState, this))
-        super.spreadTo(pLevel, pPos, pBlockState, pDirection, pFluidState);
+      if (!specialCobblegenSpread(pLevel, pPos, pBlockState, pDirection, pFluidState, this)) { super.spreadTo(pLevel, pPos, pBlockState, pDirection, pFluidState); }
     }
+
   }
+
   public static class CreataniaFlowingFluidSource extends ForgeFlowingFluid.Source {
+
     @Override
     protected void spreadTo(LevelAccessor pLevel, BlockPos pPos, BlockState pBlockState, Direction pDirection, FluidState pFluidState) {
-      if (!specialCobblegenSpread(pLevel, pPos, pBlockState, pDirection, pFluidState, this))
-        super.spreadTo(pLevel, pPos, pBlockState, pDirection, pFluidState);
+      if (!specialCobblegenSpread(pLevel, pPos, pBlockState, pDirection, pFluidState, this)) { super.spreadTo(pLevel, pPos, pBlockState, pDirection, pFluidState); }
     }
+
     public CreataniaFlowingFluidSource(Properties properties) {
       super(properties);
     }
+
   }
+
 }